requirements.txt
django-storages
google-cloud-storage 